quick-demo-version/application/requirements.txt
# Mission DB007 - Hybrid Demo Requirements
psycopg[binary]>=3.1.0
boto3>=1.26.0
colorama>=0.4.4
python-dotenv>=1.0.0